Move Async to separate file and add optional PropTypes.
