+++ src/Async.spec.js
@@ -1,7 +1,7 @@
 import "jest-dom/extend-expect"
 import React from "react"
 import { render, fireEvent, cleanup, waitForElement } from "react-testing-library"
-import Async, { createInstance } from "./"
+import Async, { createInstance } from "./Async"
 
 const abortCtrl = { abort: jest.fn() }
 window.AbortController = jest.fn().mockImplementation(() => abortCtrl)
